apply part of patch by David Evans to fix compilation on OS X 10.9
Fixed bugs:
  - https://launchpad.net/bugs/1245722

(bzr r9976)

+++ src/dom/svgimpl.cpp
@@ -81,7 +81,7 @@ static double s2d(const DOMString &s)
 typedef struct
 {
     const char *name;
-    int         type;
+    unsigned int type;
 } SVGTableEntry;
 
 
@@ -266,9 +266,9 @@ int svgInterfaceStrToEnum(const char *str)
  * Return the string corresponding to a given SVG element type enum
  * Return "unknown" if not found
  */
-const char *svgInterfaceEnumToStr(int type)
+const char *svgInterfaceEnumToStr(unsigned int type)
 {
-    if (type < 1 || type > (int)SVG_ZOOM_AND_PAN)
+    if (type < 1 || type > SVG_ZOOM_AND_PAN)
         return "unknown";
     SVGTableEntry *entry = interfaceTable;
     for (int i = 0 ; i < SVG_NR_INTERFACES ; i++)
